Update whoami router name in docker-compose.example.yml to make it easier to understand (#317)
Any router name works as long as it is consistently applied. The `nginx` name for the `whoami` container route can be a bit confusing for new users. Aligning the container and route name is similar to how Traefik generates dynamic routes, makes it easier to read the compose file and logs, and can generally help reduce bugs when extending the example.